#279 Fixed Codacy Analysis Issues and renamed some strings to account for refactoring of FileNamingStrategy
