+++ src/test/groovy/edu/ie3/datamodel/io/naming/HierarchicFileNamingStrategyTest.groovy
@@ -3,8 +3,9 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
  */
-package edu.ie3.datamodel.io.csv
+package edu.ie3.datamodel.io.naming
 
+import edu.ie3.datamodel.io.csv.DefaultDirectoryHierarchy
 import edu.ie3.datamodel.io.naming.HierarchicFileNamingStrategy
 import edu.ie3.datamodel.io.source.TimeSeriesMappingSource
 import edu.ie3.datamodel.models.BdewLoadProfile
