+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvRawGridSourceTest.groovy
@@ -743,7 +743,7 @@ class CsvRawGridSourceTest extends Specification implements CsvTestDataMeta {
 	def "The CsvRawGridSource returns an empty Optional, if one mandatory element for the RawGridElements is missing"() {
 		given: "a source pointing to malformed grid data"
 		CsvTypeSource typeSource = new CsvTypeSource(csvSep, typeFolderPath, entityPersistenceNamingStrategy)
-		source = new CsvRawGridSource(csvSep, gridFolderPath+"_malformed", entityPersistenceNamingStrategy, typeSource)
+		source = new CsvRawGridSource(csvSep, gridFolderPath + "_malformed", entityPersistenceNamingStrategy, typeSource)
 
 		when: "loading a total grid structure from file"
 		def actual = source.getGridData()
@@ -755,7 +755,7 @@ class CsvRawGridSourceTest extends Specification implements CsvTestDataMeta {
 	def "The CsvRawGridSource returns an empty Optional, if the RawGridElements contain no single element"() {
 		given: "a source pointing to malformed grid data"
 		CsvTypeSource typeSource = new CsvTypeSource(csvSep, typeFolderPath, entityPersistenceNamingStrategy)
-		source = new CsvRawGridSource(csvSep, gridFolderPath+"_empty", entityPersistenceNamingStrategy, typeSource)
+		source = new CsvRawGridSource(csvSep, gridFolderPath + "_empty", entityPersistenceNamingStrategy, typeSource)
 
 		when: "loading a total grid structure from file"
 		def actual = source.getGridData()
